requirements.txt
# Core ML Libraries
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.5.0

# Data Processing
scipy>=1.7.0
geopy>=2.2.0
pyproj>=3.2.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
folium>=0.12.0

# Configuration & Utilities
pyyaml>=6.0
tqdm>=4.62.0
click>=8.0.0
python-dotenv>=0.19.0

# Logging & Monitoring
wandb>=0.12.0
tensorboard>=2.7.0

# Testing
pytest>=6.2.0
pytest-cov>=3.0.0

# Clustering & Anomaly Detection
hdbscan>=0.8.27
pyod>=0.9.0

# Time Series Analysis
statsmodels>=0.13.0

# Geospatial Analysis
shapely>=1.8.0
geopandas>=0.10.0

# Performance
numba>=0.56.0
joblib>=1.1.0

# Development
black>=21.0.0
flake8>=4.0.0
isort>=5.9.0 